listening-comp/requirements.txt
chromadb
streamlit
# openai  # Ya no se necesita
google-generativeai>=0.3.0
youtube-transcript-api
python-dotenv